reipset: Add option to selectively SNAT to specific destinations
Currently only option for selective SNAT is packet source address.
However it might be handy to have option to SNAT for specific
destination address and port. This can be accomplished by transforming
post-fwd-snat-v4 and post-fwd-snat-v6 to list sets and adding sets to
it as elements.

Since ipset list:set does not place restriction on added as elements
sets dimensions we can add net to retain compatibility with previous
behaviour and two new sets pkt and n2n to select what to SNAT based
on packet information:

  pkt - source address, destination port, destination address
  net - source address, destination address

Note that order in which sets added to the list set makes sense so that
match stops at the first matching set. Therefore following example
will select only packets from 192.0.2.0/24 to tcp port 22 at 203.0.113.1
for SNAT to 198.51.100.1 and rest from 192.0.2.0/24 will be
SNATted to 198.51.100.2:

  ipset add post-fwd-snat-pkt-v4 192.0.2.0/24,tcp:22,203.0.113.1 \
      skbmark 0xaa000001
  ipset add post-fwd-snat-net-v4 192.0.2.0/24 \
      skbmark 0xaa000002

  ipset list post-fwd-snat-v4
  ...
  post-fwd-snat-pkt-v4
  post-fwd-snat-n2n-v4
  post-fwd-snat-net-v4

  iptables -A POSTROUTING -t nat -m mark --mark 0xaa000001 \
      -j SNAT --to-source 198.51.100.1
  iptables -A POSTROUTING -t nat -m mark --mark 0xaa000002 \
      -j SNAT --to-source 198.51.100.2

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>